pyproject.toml
[project]
name = "tgsend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.19.0",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "setuptools>=77.0.3",
]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"


[project.scripts]
tgsend = "tgsend.main:main"
